+++ modules/renderer/background_source.js
@@ -74,7 +74,13 @@ export function rendererBackgroundSource(data) {
     source.hasDescription = function() {
         var id_safe = source.id.replace(/\./g, '<TX_DOT>');
         var descriptionText = localizer.tInfo('imagery.' + id_safe + '.description', { default: _description }).text;
-        return descriptionText !== '';
+        return !!descriptionText;
+    };
+
+
+    source.descriptionText = function() {
+        var id_safe = source.id.replace(/\./g, '<TX_DOT>');
+        return t('imagery.' + id_safe + '.description', { default: _description });
     };